Move json_loads_url to utils.py
