+++ calisphere/cache_retry.py
@@ -8,16 +8,15 @@ from django.core.cache import cache
 from django.conf import settings
 
 from collections import namedtuple
-import urllib.request, urllib.error, urllib.parse
 import urllib3
 from retrying import retry
 import requests
-import pickle
-import hashlib
 import json
 import itertools
 import re
 from typing import Dict, List, Tuple
+
+from calisphere.utils import kwargs_md5
 
 urllib3.disable_warnings()
 
@@ -143,27 +142,6 @@ def SOLR(**params):
         facet_counts,
         results.get('nextCursorMark'),
     )
-
-
-# create a hash for a cache key
-def kwargs_md5(**kwargs):
-    m = hashlib.md5()
-    m.update(pickle.dumps(kwargs))
-    return m.hexdigest()
-
-
-# wrapper function for json.loads(urllib2.urlopen)
-@retry(wait_exponential_multiplier=2, stop_max_delay=10000)  # milliseconds
-def json_loads_url(url_or_req):
-    key = kwargs_md5(key='json_loads_url', url=url_or_req)
-    data = cache.get(key)
-    if not data:
-        try:
-            data = json.loads(
-                urllib.request.urlopen(url_or_req).read().decode('utf-8'))
-        except urllib.error.HTTPError:
-            data = {}
-    return data
 
 
 # dummy class for holding cached data
